[story_shell] Add containers to Story shell
Containers are pre-defined layouts of a group of surfaces,
where membership and layout are determined by properties in
surface_details.

If the story shell finds a member of a container, it determines
its layout via the semantic label containerLabel, and uses logic in
container_layout to determine the surface placement.

Current limitations:
* only the first specified layout is used
* layout is always full-screen
* no constraints for determining between layouts can be expressed
* known bug when trying to add subsequent surfaces from within container
* known bug with module pathing in peridot layer

Change-Id: Ied55a9b9a1ae6a3b7beb3979e11f0fc41448aac8