[xi-editor] Update xi_app for Dart/Flutter changes
This patch gets xi_app (the xi-editor test app) working again. It
depends on xi_core and associated fidl service being present from the
garnet layer, so deletes the (stale) implementation from topaz.

Note that this change also removes a fair amount of code for the
prototype of cross-device sync based on ledger, sadly. When we bring
that back, it will likely be useful to fetch some of that code from
the git history.

Change-Id: If059c2655cc36886b62ccf451422179aa61270f3