[todo] drop the unfinished Rust-backed TODO app.
This patch removes a skeleton of an unfinished TODO app. (Including a
backend provider in Rust, which we apparently cannot compile in topaz at
the moment.

(Side-note: the idea of splitting the app into an entity provider agent
and the UI module seems interesting. We could aim to rework this way the
other todo app in topaz: //topaz/examples/ledger/todo_list)

Change-Id: I91993a86de954a809d605f05b65797efc7e87a4a