[auth] Initial Spotify auth provider implementation.
This cl adds a skeleton impl for minting and exchanging Spotify OAuth
tokens. The implementation follows the protocol as defined in here:
https://beta.developer.spotify.com/documentation/general/
guides/authorization-guide/

Spotify OAuth doesn't support ID tokens or Firebase Tokens. As a result,
GetAppIdToken and GetAppFirebaseTokens are not supported.

Spotify doesn't provide a programmatic endpoint for revoking tokens, so
RevokeToken endpoint is unsupported as well. To manually revoke tokens,
pls follow the steps as listed here:
https://developer.spotify.com/news-stories/2016/07/25/app-ready-token-revoke

Note: Spotify requires both clientId and clientSecret for authorization
step, that should be provided as part of the api instead of hardcoding. This
requires changes to the underlying auth fidl interface which is tracked
separately.

Also added unittests for most methods except GetAppPersistentCredential
that requires a Fake web_view and other view_owner dependencies.

Change-Id: Ibc2797fae87e5e8cbbc130b4959aa19c69423ff1