[logging] Moving logging.dart to app library
The app library was using the logging library and the logging library
now needs to connect to a FIDL interface, which uses the app library.
This created a circular dependency. We could have exposed the setup of
the FIDL setup to every module that wanted to use logger but that seemed
inappropriate. Instead, we are combining the logging code into the app
library which removes the circular dependency and (after transition)
allows apps to initialize logging with the same call.

CP-31

Change-Id: Ie103d72f9cafb3af1b228e46e9828d2f58d87b15