[story_shell] Modify origin for the first Surface added to a Story
Surfaces start with a default point of origin, which is currently
immedately to the right of screen, and animate in.

For cases where some user shells may choose to animate the introduction
of Stories, prefer to have the default for the first Surface in a
Story be to appear at the final position.

Change-Id: I313b4d21a98eb975dff49235ef63c3a58cf4c50a