[logging] Adding support for logging to Zircon logger
This change allows an app to log to the zircon logger that will be
shared across all levels using a zircon socket.

It also adds support for local tags per the zircon standard for logging.

The standard dart logger cannot be extended as the only way to create
one is via factory methods. This change therefore creates an extended
logger that adds support for the new features. Apps that include this
library automatically support the new capabilities.

Modules written against the standard dart library also work as calls
into that library are redirected into the new flow after initial
processing.

This removes support for the hierarchical logger that is enabled in
the dart version.

CP-31

Change-Id: Ie877e956a10f485a53f34b02c49d2e4e8ff7c31d