[story shell] Fix layout tests
Tests were passing the full constriants to the layout files. This was
changed recently in 157878. We'll revisit this in the future.

The test utility function for validating positioned surfaces has been
updated to compensate for a flutter bug with dimensions for a rectangle
that is being tracked in issue #18169

Add tests to BUILD.gn for mondrian_story_shell.

Change-Id: I3324080a25df87cca7002b8f9c139c494f425a49