[topaz][test] Initial example IntTest code.
This code contains an initial integration test for
driver_example_mod. Once a simpler way to automate loading/running tests
is set up, this will be edited to reflect the state of things.

As of right now, this has to be run manually using existing in-tree
tools (as outlined in the `README.md`).

Change-Id: I733322be849aafe4443bcaf8f7c02e417fbc4db0