[ledger][sledge] Save document changes to Ledger.
Values now hold a ValueObserver.
The ValueObserver interface is implemented by Documents.
Whenever a Value is modified, the observer (the Document that owns the Value) forwards
the notification to the current transaction.
This allows the current transaction to track which Documents have changed.
When it is time to send the modifications to Ledger, the transaction asks the
aforementioned Documents for the Changes that need to be applied to Ledger.

Change-Id: I77232525e40c9ecaceb8164ca51a664564be927b