[ledger][sledge] Read documents back from Ledger.
When getDocument is called, a new empty document is created.
Ledger is then asked for all the KVs that start with the document's
prefix.
If KVs are found, A Change object is created using those KVs.
The Change object is then passed to the document to initialize it.

This CL fixes a bug where document's prefixes weren't appended to the
keys prior to saving the KVs in Ledger.

TEST=Ran the todo_list_sledge mod. "it works!" is printed in the logs.

Change-Id: Icef7e677cdae96d909cd44d97c67e2d190c65ecb