[ledger][sledge] Handle long keys.
Ledger have limit on key length. But Sledge CRDT's should not.
This commit provides solution for this:

Instead of storing (key, value) pair in Ledger we do modification:
We store ({hash_of_key}, {|key|}{key}{value}).

Also in this commit CRDT's tests are rewritten to not depend on
implementation of DataConverter.

Change-Id: Ia8f7cf87d35312b28f132151271e9293537e75f9