Revert "[testing] validate that unit tests are running"
This reverts commit d76e2af7dbe20c72ead898ffdd5f2df10837b707.

Reason for revert: CQ should have failed and stopped submission.

This was a test to validate that CQ is not running this (or likely any other) unit test in the topaz layer. When this problem is corrected, this test would fail. Restoring to a working state for when tests really do execute. 

Original change's description:
> [testing] validate that unit tests are running
> 
> if this passes cq, these unit tests are not being run
> 
> Change-Id: I5ec760fe349581e9c7602caf6265499aba2beed5

TBR=larrylandry@google.com,nkorsote@google.com

Change-Id: I0e0c057b5d76fc049d0981f31210b83548cf3bf0
No-Presubmit: true
No-Tree-Checks: true
No-Try: true