[auth] Add a query param to control GLIF.
The current default false should ensure we don't accidentally pick
up the "GLIF" UI style for Google sign in if our user agent changes.
GLIF is designed for modern browsers and currently does not work
well with our form factor and JS implementation. The parameter may
be changed to true for manual testing.

Change-Id: I92e3cf611a61e716f9597643f95d3697466d3d9a