[ledger][sledge] Pass random connection ID.
Some CRDT's (PosNegCounter, OrderedList) need unique ID per instance.
Such an ID's may be created per connection to Ledger, and in future
we want to be able to reuse IDs for closed connections.

In this CL random ID's are generated and pass to th CRDTs. Also
PosNegCounter was modified to work with |Uint8List| connection ID
instead of |int|;

TEST=build, test manually, pos_neg_counter_test.dart

Change-Id: I6dc3dfeaf77a30c658883e25874df774c8de809f