[ledger][sledge] Demo voting mod.
Demo app operates with list of items. Each item counts number of times it was
starred.

Each item is a Sledge Document with Schema:
{
 'title': new LastOneWinsString(),
 'votes': new IntCounter()
}

All items are stored in other Sledge document, with Schema:
{ 'items': new BytelistSet() }

Both |items| onChange and |votes| onChange streams are used. So when new
element is added list of items is rebuild. And when number of votes for
some item is changed, only that item is rebuild.

TEST=run mod, watch that it works;

Change-Id: Iaaad66b6f01145e195e97b86204bb781a29f6351