[xi] Remove xi_app window chrome
In preparation for being embedded, the xi_app module shouldn't have an
AppBar or other non-textarea UI elements.

This also adds a watermark to the background to make it easy to visually
verify that the module is running. There is some additional cleanup
(removing home_page.dart etc) that will be part of a follow-on CL.

Test: manually running the mod on the ermine shell, verifying that the
app bar isn't visible and the editor continues to function as expected.

Progress towards MS-1715

Change-Id: I2d23848eea26b38da123d617c35a832f6ed58556