[lib] Add fuchsia_driver library with example
Shows example usage of the fuchsia_driver library (the BUILD target
is not included in any packages yet and must be done manually).
`fuchsia_driver` is included in //topaz/public/lib/

The fuchsia_driver can be built either locally or run remote as a dart
application with use of the dart_runner.

Some package names were updated in order to remain more consistent with
naming conventions, so as to avoid confusion.

Packages that are purely examples (and not yet usable in CI testing)
were moved from topaz/packages/tests/examples to
topaz/packages/examples/tests file so as to avoid confusion and to
keep consistent naming.

The target
//topaz/packages/examples/tests/driver_example_mod:driver_example_mod_remote_test
has also been removed from the build, as host_tests is now run in CI,
and as it is not technically a host test (able to run in isolation
without a target device), this will break the build when included.

TEST=Running driver_example_mod_tests.dart on host and target machines.

DX-155 #done

Change-Id: I96897446df7042e5791cfd6ed4b273057a9c436b