[dart_runner] Pass List<String> not List<dynamic> to main
We were passing a List<dynamic> as the argv argument to Dart main
functions. This didn't matter in Dart 1, but in Dart 2 that makes things
like the args package's ArgParser not work.

With this change we pass a List of String rather than a List of dynamic,
containing strings to the main function.

Testing: Printed the args.runtimeType before and after, observed that I
can now use the args package in my program.

Change-Id: I00069581ea848f53a56d3edf97efea0b50ab1433