[dart][fidl] New bindings can stand alone
The new (async/await) Dart FIDL bindings used to wrap the old bindings.
Now they stand on their own.

There's a skeleton for docs but it's very bare.

The state handling has been reworked so that Bindings and Proxies share
the same state management code, there are futures for entering bound and
closed states and a stream of state changes. The error state was dropped
and it's explicitly not possible (rather than just poorly supported) to
re-bind a closed proxy or binding.

Testing: This also adds a hefty test-suite for the new bindings covering
common paths as well as lots of edge cases. It's structured as a client
and server so that it can test both the bindings and the underlying IPC
mechanisms, in case any subtle changes there break bindings semantics.

Change-Id: Id5646d1fc9b2e14699c687d9cc89a062f7656ef3