[dart][flutter][cmx] Deprecate meta/runtime
Dart/Flutter apps would like to keep their build time generated runtime
files in order to toggle between runtimes for different build
parameters. We do not have a good short term solution to address this
with component manifests, and so we are keeping separate runtime files
for now.

Complete deprecation for meta/runtime is tracked in CP-80.

shayba@: "When loading components we'd look for a manifest with a
runner. If this is missing, we'd look for a package-global runner
config."

CP-37 #progress

TESTED=ls the flutter apps and observed in the pkgfs dir it is now
deprecated_runtime
built topmost layer with release and observed pixelbook
successfully boots:
fx set x64 --args=is_debug=false out/release-x64
fx full-build
fx boot

Change-Id: I77e5f2960b3c482a0cf881bd6f9d4e8bb05af676