[message_queue] MessageQueueClient and MessageSenderClient
This CL is a followup from the FIDL message_queue.fidl changes that
change the underlying data type from String to a VMO.

This cleanup is accomodated by having everyone use MessageQueueClient
and MessageSenderClient, which converts Strings or Uint8List (which are
just bytes) to a fuchsia.mem.Buffer for convinience.

Changes include:
* Introduce MessageQueueClient (similar to the C++ version), used to
manipulating a message queue and receiving messages.
* Introduce MessageSenderClient for sending messages given a token.
* Remove MessageReceiverClient in favour of MessageQueueClient.
* zircon.SizedVmo() has a factory that takes a Uint8List.
* Update error handling on message queues to happen retroactively.

Tested: These classes are tested manually from an upper layer.

MI4-1106

Change-Id: I1556402457bed2fcad153632c5d5c7442724f0f8