[ermine] Rebuild ermine as a configuration of 'tiles'
This removes the old ermine user shell implementation and instead
defines the 'ermine' product configuration to boot up to tiles with a
single tile containing the program 'term'. More tiles can be added or
removed with the 'tiles_ctl' command.  The ermine configuration also
contains dashboard and noodles_view to have pretty things to look at.

The default configuration for the topaz layer also boots to ermine,
with the addition of all of the production packages within topaz.

Test: boot topaz/packages/default and topaz/packages/products/ermine
Change-Id: Id6e627422a4399053faacbf020cdce6d23c5e165