[ledger][sledge] Include Sledge version in PageIds.
The goal of this CL is to avoid crashes when the format of what is
stored in Ledger changes.
The goal of this CL is *not* to implement a way for us to migrate data
between Sledge versions.

The way this goal is implemented is by having the version of Sledge be part of
the SledgePageIds.
Also in this CL: the user-provided page names are hashed, and the hash
function in Sledge is changed from sha1 to sha256.

TEST=sledge_page_id_test.dart

Change-Id: I05b2d38eb92c8a621fe0a38a1742c4266c89fbea