[ledger][sledge] Random synchronizations.
Allows to enable a random synchronization.

With .setRandomSynchronizations one can specify an expected number of
additional synchronization nodes after execution of each node of the
computational graph.

Those synchronization nodes are generated during runtime. So on
different executions they may appear in a different positions.

Those nodes are not a part of a computational graph.

However those nodes are specified in a SingleOrderTestFailure message.
And when order is restoring from the list of nodeIds, the same set of
generated synchronization nodes is added. It's possible to do because
synchronization nodes are fully described by they nodeId, which have
a form.

This CL also change the format of all nodeIds. New format is:
'typePrefix[-id1[_id2]]-uniquePart'
Where typePrefix is:
    - 'n' for manually generated nodes
    - 'm' for modification nodes
    - 's' for synchronization nodes
    - 'g' for generated synchronization nodes

TEST=dart_sledge_tests

Change-Id: Ic37cac72315d74f0a984ce08c954f9300341c58c