[media] handle normalizedSeek before load is complete
This CL makes the media player controllers handle normalizedSeek
while the content is loading. It also fixes a couple of bugs.

AudioPlayerController and MediaPlayerController have two flavors of
seek: 'seek', 'normalizedSeek'. normalizedSeek accepts a value in
the range 0..1. It needs a valid duration in order to know how to
seek (because the FIDL player doesn't do normalized seek). It used
to do nothing as long as the duration wasn't known. Now it caches
the request and completes it when duration is known.

There were also two bug fixes:

1) The 'loading' property wasn't set to true immediately on open.
The transition would come later (if at all) when the player reported
its new status. This made the player appear ready when transitioning
from one URL to another.

2) When transitioning from on URL to another (as opposed to starting
a new player), file-protocol URLs were not properly handled.

TEST: manually tested with specially-rigged media_player_flutter app

MTWN-156 #done
MTWN-157 #done

Change-Id: I9e97eb46f209aa01387d6a4479f099d8eb5474c7