[xi] Add interface and impl for core and view
This separates the interface for the core from the interface for the
view, and provides base impls for both. In subsequent work, there will
be another impl for the view interface that supports agent-managed
views.

This CL is only adding these interfaces; it does not change the current
code to use them.

Progress on MS-1903
Test: No functional change

Change-Id: Ib628ce3fc1c36b1b5f50ed3c5db9ffed2efb5da9