[SetUI] Module-based roster and action handling.
This changelist introduces the SetUI's  action functionality. The
action implementions are how steps defined in the syllabus are
expressed. The module-based action handling is the first
implementation. The conductor (not yet implemented) will use the
module action handlers to launch the corresponding modules defined in
the roster. On the other side, client modules will use the module
action helper to send back results.

Test: fx run-host-tests lib_setui_module_test
Change-Id: I1d6dc5bd89ef723769a9657763f1357f9f7c66a3