[ledger][sledge] Optimize operations on Uint8Lists.
getUint8List(Prefix|Suffix) would create a copy of the contents of
the given argument. This CL adds getSublistView method that replaces
those 2 and and also updates the implementation to use an unmodifiable
view on the argument instead.

Also in getUint8ListFromString, avoid making an unnecessary copy of a list.

Test: existing tests still pass

Change-Id: I0aded6a312ae9b60554639436f935b524b8a3f2e