[ledger][sledge] Store a boolean to mark Document as existing.
This CL introduces the concept of private and public fields in
Documents.
The public fields are the ones defined in the Schema.
The private fields hold metadata about the Document.

In this CL, the only private field is a LastOneWins bool that stores
whether the Document exists.
As soon as the Document is created, the field is set to True. This
`write` will be propagated to Ledger, guaranteeing that the Document is
backed by at least one key-value in Ledger.

This CL also adds the |documentExists| method which returns whether a
Document exists.
The implementation of this method simply checks whether a key-value
exists for the given Document.
Because of the aforementioned LastOneWins bool, a key-value is guaranted
to be stored in Ledger for every Document that has been created.

Lastly, SledgeTestingMod was currently failing because of a previous
change. This CL fixes SledgeTestingMod, and adds tests for
|documentExists|.

TEST=Added test in sledge_testing_mod.

Change-Id: Ia1ea5296a66f3963eb587d1254392ef3e0cb7a06