[ermine] Initial version of ermine
Ermine is a Fuchsia User Shell intended to provide developers with a
lovable environment for testing mods on Fuchsia.

This initial version implements the same approach for view hosting,
layout and control as //garnet/bin/developer/[tiles|tiles_ctl]. For
simplicity it re-uses
//garnet/public/fid/fuchsia.developer.tiles/tiles.fidl.

This change contains three major parts:

//topaz/bin/user_shell/ermine_user_shell

This is the binary that is launched by device_runner and acts as a user
shell. It draws a background color, takes care of launching new stories
via the story provider and lays out the views that display each story.

//topaz/bin/ermine_ctl

This is a binary that knows how to ask ermine to add, remove and list
the stories it is hosting via the fuchsia.developer.tiles interface.

//topaz/packages

This is a set of configuration changes needed to get ermine launched as
a user shell.

There are a few other changes supporting Rust development in Topaz.

This change also deletes the previous program named ermine, which was
not being used even before this change.

Testing: built, booted, used ermine_ctl to add, remove and list stories.

Change-Id: I2fd657188ca7acb3dd373aa97058b1aa3b1c0ba5