[fidl][dart] "String" is a reserved name, escape with $
"String" needs to be a reserved name because we use the dart:core String
type in FIDL generated bindings and if you name a FIDL type "String"
things will be ugly.

Instead of adding a trailing "_" to reserved names, add a trailing "$"
since a trailing "_" breaks the camel-case type name requirements
imposed by the dart linter. Unfortunately "$" is the treated specially
in Dart strings so some need to be raw strings.

TEST=built after adding "//zircon/public/fidl/fuchsia-net:fuchsia-net"
as a dep to a Dart target.

FIDL-257 #done

Change-Id: I14021918bd491347af0d34c2d3b28d4dd12fe76a