[logging] Fix log severity to use Int32 instead of Uint32
Test: Tested manually. Existing unit test missed due to internals.

The previous code called setUint32 on ByteData passing an Int32 instead
of calling setInt32. Since both were the same length, the internal code
of ByteData processed this with the same result either way. However,
the code is technically wrong. This fixes that issue. The output byte
stream does not change.

Change-Id: Ied7c26001f7553fa01cfec4e489810f1772f1852