[flutter/dart] Remove 'aot' flag from {dart,flutter}_app template
This is another way to control whether the runtime execution mode for a
Dart or Flutter program, in addition to picking the mode with the
{dart,flutter}_aot_app templates. The 'aot' flag is unused except for
tests and armadillo_user_shell. This removes the flag and updates
armadillo_user_shell to use the same mechanism as the rest of the tree
to make simplify the system for future cleanups.

Test: Build in debug & release and confirm armadillo_user_shell is in
the expected runtime mode.

Change-Id: I61b4dc48837b2d39e3537ad7bb0e602f564f343e