[SetUI] Widget-based action introduction.
This changelist introduces actions handled by widgets. This can be
used as either the client side of a module based setup flow or
internally for a self-contained flutter app flow.

This change also revises how action-related code is structured.
Blueprints are now registered with a generic Roster class. These
Blueprints are used to generate an Action (roughly replacing
ActionHandler), which is a running instance. ActionClients base
implementations for user facing components that communicate back to
the "server-side" Action components.

Test: fx run-host-tests lib_setui_flutter_test
Test: fx run-host-tests lib_setui_common_test
Test: fx run-host-tests lib_setui_module_test
Change-Id: I950e7bfed1d85cfc3c1f8465269208f94ebf2849