[SetUI] Couple Actions to Blueprints.
This change makes the base Action class take a blueprint as a generic
type. This allows for the Blueprint to be part of the base class,
while also being accessible to the Action owner. This changelist also
improves the debugging information present in the Blueprint, allowing
for better Action identification.

Test: fx run-host-tests lib_setui_flutter_test
Test: fx run-host-tests lib_setui_common_test
Test: fx run-host-tests lib_setui_module_test
Change-Id: I30b519abccaeffcf55b695b528e6348958089d29