[xi] Add tabbed multi-view support to xi_mod
This adds very basic multi-view support to xi_mod. Tabs can be added and
removed, and each tab is backed by its own xi-core view.

This uses the Material TabBar and TabBarView classes, which were not
particularly intended for this use-case, but serve to demonstrate that
multiple views is possible.

Test: None
Change-Id: I3558c498152091a1c699799a0a4a3034c9bc192b