[auth] Add Chromium.web integration.
This is the second step in migration from Webkit to Chromium. It will
require the chromuim interface to land in Fuchsia before commit and the
Chromium implementation to land in Fuchsia before it wil run.

Further work is likely to be required passing an appropriate service
directory to the Chromium context.

AUTH-63

Tested: None yet, AUTH-70 tracks creation of a test that will allow
manual verification.

Change-Id: Iaae3042da065112c47a289be8279be8f24aa3c34