[module_driver] Add IntentHandler to module driver
This adds an IntentHandlerImpl which provides an IntentHandler in the
module's outgoing services. The framework connects to this service each
time a new intent is to be handled by the module. A new connection is
made for each new intent.

TESTED=Manually by verifying that the module driver logs are printed
when a module is started.

MI4-1249 #done

Change-Id: I096f6db27cdb6f45b1736cda919eb22e34cfd9d5