[module_driver] Add onHandleIntent to ModuleDriver
This allows modules to provide intent handling functions, while letting
the module driver buffer incoming intents to avoid forcing the module to
register the callback prior to the first run loop.

This also adds an IntentParameters class which gives modules access to
intent parameters in a unified way (i.e. regardless if they are links,
entities, or JSON). The module can get parameter data directly by
calling getParameter, or it can watch a parameter with watchParameter
if the module expects the value to change.

This IntentParameters struct is an intermediate state, and once all
intent parameters are delivered as entities the signature of the
OnHandleIntent can become:

void Function(String action, Map<String, Entity> parameters);

TESTED=Manually.

Change-Id: Ia2f90b95c6ccfb167228b8b2425227d85906c58d