[dart][flutter] Add dart_force_product
This CL adds a "dart_force_product" flag to the build that forces all
Dart and Flutter apps to use a specific product configuration regardless
of how they are declared. We can use this flag when building certain
product configurations to minimize storage footprint.

Test: Build with dart_force_product=true and see that all Dart and
Flutter apps are built in JIT mode.

Change-Id: Ia05259c2dc18dd55d925403c5972324775d99963