[flutter] Reconnect refocused IMEs in PlatformView
Previously, when a PlatformView lost focus, it would disconnect Flutter
from the IME. When the PlatformView regained focus, Flutter would think
the textarea was still connected, and so not request a new IME
connection. This meant input was disabled until textarea focus was
switched in the Flutter app.

After this change, PlatformView will save the last state of the text
field, and when it regains focus (if Flutter still wants text input)
will reconnect the IME.

This change also updates the deprecated Hide and Show calls on the IME
to HideKeyboard and ShowKeyboard on the ImeService.

TEXT-13 #done

SCN-928 #comment tracks Scenic telling the IME about focus, which will
allow us to remove the IME disconnection code entirely; disconnection
can happen entirely within the ImeService instead.

Test: manually tested change works with small example program; there do
not appear to be automated tests for PlatformView

Change-Id: Id85a3f4d4671231da832aff723abcf3e8df87004