[fidl][dart] Generate constructors to clone structs
FIDL structs are immutable in Dart. This is convenient for all sorts of
reasons but occasionally it's useful to be able to make a copy of a
struct with some of the fields changed.

This adds one or two constructors to every generated struct that can be
used to create a copy of an existing struct with some fields changed or
unset.

All structs get a new `clone()` constructor that takes as its first
argument an existing struct of the same type and then named arguments
for each struct member. It returns a copy of the existing struct with
field values specified in the named arguments overriding ones in the
existing struct.

Unfortunately this won't allow a nullable field to be set to null as
there's no way to tell that a null value was passed rather than the
argument not being passed. So structs with any nullable fields will have
a `cloneWithout()` method taking an existing struct and named bool
arguments for each field. A copy will be returned with any specified
fields nulled.

TEST=fx shell runtests /pkgfs/packages/fidl_bindings_test/0/test

FIDL-278 #done

Change-Id: I40f0e2e604e070a71d9a49e6e051605dc282ad86