[flutter][input] Hook up Scenic input events.
Most of this CL is dedicated to moving the SessionListener to
PlatformView (from SessionConnection) and simplifying the result. In
particular, it allows Flutter to receive Scenic input events independent
of the View system it is compiled for. It enables a smoother transition
to Scenic inputs without blocking on a global views v2 transition.

SCN-840 #comment

SCN-976 #done

TEST=$ set_root_view simple_flutter # no regression
$ present_view simple_flutter # new input stack works
manual testing with full stack

Change-Id: Idbc56fd60d32c3e6f1ef805b6af594e7ad293c1f