[story_shell] (bug fix) Unregister channel listener when not maximized
We register a listener for listening for key commands in Story Shell.

We register when we are maximized. The CL to add this forgot to
unregister when this is no longer true which (I think) lead to bad
behaviour, e.g. the ChannelReader issues referenced in MI4-1332

test: manual test - without the fix I could replicate the channel
listener error. With the fix I failed to replicate channel listener
error.

Change-Id: Id0afebfc963d8b104f18f49a5bfc4fdd9569c2e3