[components] Add correct program.data to cmx
Now that we read program.data for flutter and dart apps, we declare them
explicitly in cmx files. This cuts down on the number of inferred
default component names. Also, it is just plain wrong to have these apps
specify a bin/app.

This is a query for gn rules that include cmx files in topaz:
package:^fuchsia-topaz f:gn cmx
We then manually inspect each build rule that includes a cmx, to see if
they are a flutter or dart app, and appropriately add the correct entry
in the cmx.

TESTED=CQ

CP-19 #comment

Change-Id: I58ad6ab439ade890b3af0f189b6fb0e96e0b6ce7