[SetUI] Introduce settings library.
The SetUI Settings library provides an abstractions for client
implementations to access settings. The library is based on an
MVC interaction, implementing the model and controller. The
controllers interact with settings through adapters.

This changelist also implements a time zone adapter that
directly interacts with the time zone fidl service.

Test: fx run-host-tests lib_setui_settings_test
Change-Id: I6cbe2ad3c0edc7fac2544756b34e24fd1c0e8973