[components][flutter] Add correct program.data to cmx
Now that we read program.data for flutter and dart apps, we declare them
explicitly in cmx files. This cuts down on the number of inferred
default component names. Also, it is just plain wrong to have these apps
specify a bin/app.

data = /pkg/data for the cmx's that we missed in
https://fuchsia-review.googlesource.com/c/topaz/+/193280

TESTED=CQ

CP-19 #comment

Change-Id: I83276e6c46ffa61b881e111253b5c74565bfb305