[module_driver] introduce #startSync()
and deprecate #start(), since there is really no reason
to await for anything inside the start method. If anything
fails in start it means the framework is down or in some
inconsistent state.

This change should surface errors much quicker, it will
prevent users from swallowing the an error by simply
logging a warning and proceeding.

MS-2062 #done
Test: Manually by running a mod with driver.startSync();

Change-Id: I493a690ac0630414ef44cc47faca83fc3e109530