[story_shell] Add placeholder color to surfaces
This change provides a place for module authors to specify a color
they would like to use in a placeholder that is shown between their
module being started and their module rendering pixels.

This is to facilitate the current plan of record re placeholders:
Story Shell will show a widget of specified color to finesse over
mod loading time. The fallback color is white.

examples/mondrian_test/manifest.json has a canonical example of
specifying placeholder_color in a module.manifest.

Placeholder color is specified on the module_manifest using the
placeholder_color parameter, and accepts HTML color values in the format
'#RRGGBB'.

When a module is launched, StoryShell shows an empty widget of either
the placeholder_color or white, and in lieu of Scenic signalling,
switches to the ChildView after _kPlaceholderDuration defined in
models/surface/surface_transition.dart.

Test: manual - both example_manual_relationships and mondrian_test have
been modified to have placeholder colors of '#00FF00' and '#FF00000'
respectively, and can be launched to check the behaviour.
In the longer term we will want a more comprehensive StoryShell test.

SY-483: DONE
MS-1510: DONE
Change-Id: I4ac2d690b7a896086d1678d87473f489579309f7