[dart][flutter] Add dart_force_product
This CL adds a "dart_force_product" flag to the build that forces all
Dart and Flutter apps to use a specific product configuration regardless
of how they are declared. We can use this flag when building certain
product configurations to minimize storage footprint.

Test: Build with dart_force_product=true and see that all Dart and
Flutter apps are built in JIT mode.

Change-Id: Ibc79b623bd87b85ee214db5eeeb595379cc1d8fd