[flutter] Adds a space_dart parameter to flutter_jit_app()
The parameter defaults to false. When it is true, bytecode
is included in the kernel files for the app, and the
flutter_runner passes the --enable_interpreter flag to the
Dart VM.

Due to temporary incompatibilities between core-jit
snapshots generated with and without the interpreter
enabled, the core snapshot for the JIT runner is changed
from a core-jit snapshot to a regular core snapshot. This
will be changed back to a core-jit snapshot when the
temporary incompatibilities are addressed.

TEST=CQ,hello_space_dart
Change-Id: I3c6489e6fd033cf59b5ef809ebdd0f178711774a