[jiri] Increase history depth to 5 for third party repos.
This will reduce the probability of a jiri update failing during a
release build. Currently those builds are failing if they refer to a
commit that is anything other than the most recent commit in that repo.

This is a temporary fix. See bug for context and long term plan.

Test: not needed
Bug: INTK-485 #comment
Change-Id: Ief0ef2d25b8d8d7f71f272b71c7d04ac520c5ee4