[repl] REPL server
The REPL server is a test process that connects to multiple components
and interacts with them.

This is phase one, with a shell script that runs a device_runner command
based on driver_example_mod_test. In future CLs, the REPL server will get
its own test binary to operate from.

Test: Added a debugger() line to driver_example_mod_test.dart and
followed the instructions in tools/repl/lib/main.dart

Change-Id: Ib44570d46e4793f436ecac82bce96152e933b694