[story shell] Handles surfaces that aren't related
Surfaces that are added to the story without a parent are connected to
the root of the tree. When laying these out, they would not stack on top
of each other because they were previously not reachable when
determining layout. Instead of looking for related surfaces from the
current surface, use the root of the tree instead.

TEST=manual

SY-611 #DONE

Change-Id: I8d0fa389424879142b27da5d9d2523b60e1f7acd