[flutter] Reuse large enough |VulkanSurface|s in |VulkanSurfacePool|
Modify |VulkanSurfacePool| to allow available |VulkanSurfaces| of
sufficient |VkDeviceMemory| size to be used as new surfaces that do not
perfectly match the size of the existing |VkImage|.

This is accomplished by allowing |VulkanSurfaces| to swap their current
|VkImage| to a new one, without reallocating its |VkDeviceMemory|.  This
is important, as |vk().AllocateMemory| heavily dominates the GPU
workload of Flutter apps during resizes.

In the case of WoD, there's about 10 megabytes at stake when it comes to
shrinking oversized surfaces at some point.

On a WoD resize down, average GPURasterizer::Draw (in the subsequent
750ms after the first frame after the touch event) wall time went
from ~8.2ms to ~4.9ms.

TODO (after submitting this CL):

This change as is significantly improves graphics performance in the
case where a window resizes down, as there will in general surfaces of
sufficient size available to reuse.  In the case where we want to resize
up however, we will in general miss the cache.  We need to send a signal
from somewhere to grow the surfaces in the cache up to a suitable size
before the resize up animation starts.

Test: CQ, manually resizing WoD+perspective a bunch
Change-Id: I81a7759ab18e706d6042987759c5df68d49802e8