[ledger][sledge] Correctly rollback document creation
When a document is created in a transaction that is then rolled back,
the document object still exists in sledge's internal state, and might
be leaked from the aborted transaction. Such documents are now
considered invalid, and operations on them fail. When getting the
document again in a new transaction, we ensure that the document is
marked as existing in ledger, and make the document object valid again.

LE-588 #done
tested: new test in document_test.dart

Change-Id: I62ff1467e1988d69034e7a373c91420421527f91